requirements.txt
cheese
coolname
datasets
environment
evaluation
gradio
hydra-core
numpy
omegaconf
pandas
pynvml
rouge
sacrebleu
tqdm
transformers
wandb
scikit-learn
accelerate
evaluate
optimum
sentence-transformers
openai
scipy
plotly
